requirements.txt
autogen-agentchat
autogen-ext[openai,mcp]
rich
python-frontmatter
autogen-contextplus
chevron
requests-cache
